Updated method AddConferenceListOptions
Updated line 203 to check if options.Status IsNullOrEmpty to reflect changes made to the ConferenceListRequest object property "Status" from int to string.  Twilio uses a string for the Status option and expects a status string of "init", "in-progress", or "completed" which cannot be passed to Twilio in the form of an Int type.