Removed 3 duplicated properties (1 having a type mismatch that caused an exception), added 3 undocumented properties.
WebApi's JQueryMvcFormUrlEncodedFormatter was throwing an
System.Reflection.AmbiguousMatchException because RecordingDuration was
defined in this class as a float and in the base class VoiceRequest as a
string, this caused the controller StatusCallbackRequest parameter to be
passed as null.  Removed this and RecordingUrl, RecordingSid as they
were also defined in VoiceRequest.  Added Called, Caller & Duration
properties that are not documented at
http://www.twilio.com/docs/api/twiml/twilio_request#asynchronous but are
coming through from Twilio.