Reconfigure TwiMLResult to prevent XXE
TwiMLResult provides an additional (unused by the library)
constructor that accepts a string value and will use
`XDocument.Parse()` to turn it into an XML Document.

Older versions of .Net allow DOCTYPE defined ENTITYs by default.
Careless use of this class (passing in unsanitized user data) is
vulnerable to an XXE Attack.

This commit changes the default behavior to protect against this
threat with associated positive and negative tests.

For more information about this type of threat please see:
https://www.owasp.org/index.php/XML_External_Entity_(XXE)_Processing

For .Net specific information and the reference solution this
commit is based on please see:
https://msdn.microsoft.com/en-us/magazine/ee335713.aspx